#234 refactor: ディレクトリ削除
