+++ src/app/tag/[tagId]/page.tsx
@@ -1,6 +1,6 @@
 import { getList } from '@/libs/microcms';
 import { LIMIT } from '@/constants';
-import Pagination from '@/components/Layouts/Pagination';
+import Pagination from '@/components/Common/Pagination';
 import ArticleList from '@/components/ArticleLists/ArticleList';
 import TopSidebar from '@/components/Sidebars/TopSidebar';
 import Display from '@/components/Adsense/Display';
