--- src/components/Categories/CategoryList/index.module.css
@@ -1,6 +1,0 @@
-.categories {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin: 8px 0;
-}
