--- src/components/Categories/CategoryListItem/index.module.css
@@ -1,6 +1,0 @@
-.category {
-  border-radius: var(--border-radius);
-  font-weight: bold;
-  font-size: 15px !important;
-  white-space: nowrap;
-}
