--- src/components/Layouts/Pagination/index.module.css
@@ -1,30 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 24px;
-  margin-top: 24px;
-}
-.list {
-  margin: 0 4px;
-}
-.item {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 36px;
-  height: 36px;
-  border-radius: var(--border-radius);
-}
-
-@media (max-width: 768px) {
-  .container {
-    margin-top: 0;
-  }
-}
-
-@media (max-width: 640px) {
-  .container {
-    margin-bottom: 1.75rem;
-  }
-}
