--- src/components/Tags/TagList/index.module.css
@@ -1,6 +1,0 @@
-.tags {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin: 8px 0;
-}
